+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Add x265 in the `cppbuild.sh` script for FFmpeg, thanks to Mark Bolstad
  * Upgrade presets for FFmpeg 2.5.4, OpenSSL 1.0.2
  * Switch from `IntPointer` to `BoolPointer` for the `BOOL*` pointer type of Leptonica ([issue #36](https://github.com/bytedeco/javacpp-presets/issues/36))
  * Add `preload` for `gif`, `jpeg`, `png`, `tiff`, and `webp` libraries in presets for Leptonica ([issue #36](https://github.com/bytedeco/javacpp-presets/issues/36))